Changed IntResolver to UIntResolver. It now validates if argument is negative so we also have NumberAssertion and NegativeNumberException back

+++ Backend/Console/Commands/Resolver/Pipelines/CurrencyUpdateResolver.cs
@@ -1,4 +1,5 @@
 ﻿using Console.Commands.Domains.Arguments;
+using Console.Commands.Resolver.Exceptions;
 using IdelPog.Common.Enums;
 
 namespace Console.Commands.Resolver.Pipelines
@@ -6,10 +7,10 @@ namespace Console.Commands.Resolver.Pipelines
     public class CurrencyUpdateResolver : IArgumentResolverPipeline<CurrencyUpdateArguments>
     {
         private readonly IArgumentResolver<ActionType> _actionTypeResolver;
-        private readonly IArgumentResolver<int> _intResolver;
+        private readonly IArgumentResolver<uint> _intResolver;
         private readonly IArgumentResolver<CurrencyType> _currencyTypeResolver;
 
-        public CurrencyUpdateResolver(IArgumentResolver<ActionType> actionTypeResolver, IArgumentResolver<int> intResolver,
+        public CurrencyUpdateResolver(IArgumentResolver<ActionType> actionTypeResolver, IArgumentResolver<uint> intResolver,
             IArgumentResolver<CurrencyType> currencyTypeResolver)
         {
             _actionTypeResolver = actionTypeResolver;
@@ -20,13 +21,13 @@ namespace Console.Commands.Resolver.Pipelines
         public CurrencyUpdateArguments Resolve(ReadOnlySpan<string> arguments)
         {
             ActionType actionType = _actionTypeResolver.Resolve(arguments[0]);
-            int amount = _intResolver.Resolve(arguments[1]);
+            uint amount = _intResolver.Resolve(arguments[1]);
             CurrencyType currencyType = _currencyTypeResolver.Resolve(arguments[2]);
 
             return new CurrencyUpdateArguments
             {
                 ActionType = actionType,
-                Amount = (uint) amount,
+                Amount = amount,
                 CurrencyType = currencyType
             };
         }
